+++ idea/testData/refactoring/pullUp/k2k/reformatModifierList.kt
@@ -1,0 +1,8 @@
+annotation class Concat
+
+open class A
+
+class <caret>Abstraction : A() {
+    // INFO: {"checked": "true", "toAbstract": "true"}
+    @Concat var extraction = 0
+}